requirements.txt
appdirs==1.4.4
asgiref==3.3.4
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.2
django-crispy-forms==1.11.2
django-filter==2.4.0
django-tinymce==3.3.0
djangorestframework==3.12.4
Eel==0.14.0
filelock==3.0.12
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
Markdown==3.3.4
oauthlib==3.1.0
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
SpeechRecognition==3.8.1
sqlparse==0.4.1
urllib3==1.26.4
virtualenv==20.4.4
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
